+++ .jazzy.yaml
@@ -2,8 +2,28 @@ author: Kraig Spear
 copyright: 'Copyright © 2020 DornerWorks. All rights reserved.'
 documentation: guides/*.md
 
-#custom_categories:
-#- name: HomeKit
-#  children:
-#    - RoomProtocol
-#    - Room
+custom_categories:
+- name: RoomsView
+  children:
+  - RoomsView
+  - RoomsViewModel
+  - FilterSortButton
+  - TitleView
+- name: RoomLightsView
+  children:
+  - RoomLightsView
+  - RoomLightsViewModel
+  - RoomViewStyleModifier
+- name: Database
+  children:
+  - RoomDataAccessible
+  - RoomAccessor
+- name: General
+  children:
+  - BaseOperation
+  - HapticFeedbackProtocol
+  - HapticFeedback
+  - RefreshNotificationProtocol
+  - RefreshNotification
+  - URLOpenable
+  - URLOpener
