+++ HomeKitLights/General/HapticFeedback.swift
@@ -10,7 +10,7 @@ import UIKit
 
 /// Provides Haptic feedback
 protocol HapticFeedbackProtocol {
-    /// Impact has occurred
+    /// This method tells the generator that an impact has occurred
     func impactOccurred()
 }
 
@@ -23,10 +23,12 @@ final class HapticFeedbackMock: HapticFeedbackProtocol {
 final class HapticFeedback: HapticFeedbackProtocol {
     private let feedGenerator = UIImpactFeedbackGenerator(style: .medium)
 
+    /// Static instance, only once instance should be created
     static var sharedHapticFeedback = HapticFeedback()
 
     private init() {}
 
+    /// This method tells the generator that an impact has occurred
     func impactOccurred() {
         feedGenerator.impactOccurred()
     }
