+++ HomeKitLights/HomeKit/Models/Accessory.swift
@@ -8,13 +8,22 @@
 
 import Foundation
 
-/// Represents an Accessory from HomeKit
-struct Accessory: Identifiable, Hashable {
+/// Struc representation of an `HMAccessory` from HomeKit
+/// - SeeAlso: `HMAccessory`
+struct Accessory: Identifiable {
+    /// Name of accessory
     let name: String
-    let id: UUID
-    let isOn: Bool
-    let brightness: Int
 
+    /// UUID identifying this Accessory
+    let id: UUID
+    /// Is this Accessory on/off
+    let isOn: Bool
+
+    /// Brightness level for this Accessory
+    let brightness: Int
+}
+
+extension Accessory: Hashable {
     func hash(into hasher: inout Hasher) {
         hasher.combine(id)
     }
